Update mechme.java
Rename variables Lucas, Aiden, and armL. In addition to this, I added a comment to the potato variable to make it more clear what the value does.

In case the varibles need to be changed in the future, or you wish to know what I changed each value to, I will specify that below:

armL var name change:
"armL"has been changed to just "arm" as there is only one arm.

lucas var name cange:
"lucas" has been changed to "sholderPosition" so that it can better convey what it is used for.

aiden var name change:
"aiden" has been changed to "elbowJoint" to better convey what the  variable does.